requirements.txt
boto3>=1.39.9
botocore>=1.39.9
numpy>=1.24.0
sentence-transformers>=2.2.0
Pillow>=10.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
requests>=2.31.0
pypdf>=3.0.0 